requirements.txt
python-telegram-bot==21.*
httpx
python-dotenv
